🚚 Renamed `NavBar` layout component and added to index
